+++ Shell.Controls/LiveTilesLayoutControl.xaml
@@ -1,0 +1,83 @@
+﻿<UserControl
+    xmlns:tiles="using:NotificationsVisualizerLibrary" 
+    xmlns:tilesAccessLibrary="using:Shell.LiveTilesAccessLibrary" 
+    x:Class="Shell.Controls.LiveTilesLayoutControl"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:local="using:Shell.Controls"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    xmlns:media="using:Windows.UI.Xaml.Media"
+    mc:Ignorable="d"
+    d:DesignHeight="300"
+    d:DesignWidth="400">
+
+    <Grid>
+        <ScrollViewer x:Name="RootScrollViewer">
+            <GridView
+                x:Name="LiveTiles"
+                ItemsSource="{x:Bind ItemsSource, Mode=TwoWay}"
+                AllowDrop="True"
+                CanReorderItems="True"
+                SelectionChanged="LiveTiles_SelectionChanged"
+                SelectionMode="Single">
+                <GridView.ItemsPanel>
+                    <ItemsPanelTemplate>
+                        <VariableSizedWrapGrid
+                            x:Name="LiveTilesLayout"
+                            MaximumRowsOrColumns="8"
+                            ItemHeight="92"
+                            ItemWidth="92"
+                            AllowDrop="True"
+                            Orientation="Horizontal"
+                            HorizontalAlignment="Center"
+                            HorizontalChildrenAlignment="Stretch"
+                            VerticalChildrenAlignment="Stretch"
+                            Loaded="LiveTilesLayout_Loaded"
+                        />
+                    </ItemsPanelTemplate>
+                </GridView.ItemsPanel>
+
+                <GridView.ItemTemplate>
+                    <DataTemplate x:DataType="tilesAccessLibrary:TileModel">
+                        <Grid
+                            x:Name="LiveTile"
+                            VariableSizedWrapGrid.RowSpan="{x:Bind RowSpan}"
+                            VariableSizedWrapGrid.ColumnSpan="{x:Bind ColumnSpan}"
+                            Tapped="LiveTile_Tapped"
+                            Loaded="LiveTileRoot_Loaded"
+                            CornerRadius="4">
+                            <Grid.Background>
+                                <media:AcrylicBrush
+                                    BackgroundSource="HostBackdrop"
+                                    TintOpacity="0.25"
+                                    TintColor="{ThemeResource SystemAccentColorLight2}"
+                                    FallbackColor="{ThemeResource SystemAccentColor}" />
+                            </Grid.Background>
+
+                            <Grid
+                                x:Name="TileWrapper"
+                                Background="{x:Bind Logo}">
+                                <tiles:PreviewTile
+                                    Content="{x:Bind LiveTile}"
+                                    Loaded="LiveTile_Loaded"/>
+                            </Grid>
+
+                            <Grid.ContextFlyout>
+                                <MenuFlyout>
+                                    <MenuFlyoutItem Text="Size"/>
+                                    <MenuFlyoutSeparator/>
+                                    <MenuFlyoutItem x:Name="SmallOpt" Text="Small" Click="LiveTileContext_Click"/>
+                                    <MenuFlyoutItem x:Name="MediumOpt" Text="Medium" Click="LiveTileContext_Click"/>
+                                    <MenuFlyoutItem x:Name="WideOpt" Text="Wide" Click="LiveTileContext_Click"/>
+                                    <MenuFlyoutItem x:Name="LargeOpt" Text="Large" Click="LiveTileContext_Click"/>
+                                    <MenuFlyoutSeparator/>
+                                </MenuFlyout>
+                            </Grid.ContextFlyout>
+                        </Grid>
+                    </DataTemplate>
+                </GridView.ItemTemplate>
+            </GridView>
+        </ScrollViewer>
+    </Grid>
+</UserControl>
