+++ Shell.Controls/Shell.Controls.csproj
@@ -120,6 +120,9 @@
     <RestoreProjectStyle>PackageReference</RestoreProjectStyle>
   </PropertyGroup>
   <ItemGroup>
+    <Compile Include="LiveTilesLayoutControl.xaml.cs">
+      <DependentUpon>LiveTilesLayoutControl.xaml</DependentUpon>
+    </Compile>
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="StartScreenControl.xaml.cs">
       <DependentUpon>StartScreenControl.xaml</DependentUpon>
@@ -132,6 +135,10 @@
     </PackageReference>
   </ItemGroup>
   <ItemGroup>
+    <Page Include="LiveTilesLayoutControl.xaml">
+      <SubType>Designer</SubType>
+      <Generator>MSBuild:Compile</Generator>
+    </Page>
     <Page Include="StartScreenControl.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
